Remove secure dependency on lib library
