+++ nano/lib/work.hpp
@@ -37,7 +37,7 @@ public:
 	void generate (nano::uint256_union const &, std::function<void(boost::optional<uint64_t> const &)>, uint64_t);
 	uint64_t generate (nano::uint256_union const &);
 	uint64_t generate (nano::uint256_union const &, uint64_t);
-	nano::network_params network_params;
+	nano::network_constants network_constants;
 	std::atomic<int> ticket;
 	bool done;
 	std::vector<boost::thread> threads;
