+++ nano/node/node.cpp
@@ -737,7 +737,7 @@ void nano::vote_processor::vote (std::shared_ptr<nano::vote> vote_a, std::shared
 		/* Random early delection levels
 		Always process votes for test network (process = true)
 		Stop processing with max 144 * 1024 votes */
-		if (!node.network_params.is_test_network ())
+		if (!node.network_params.network.is_test_network ())
 		{
 			// Level 0 (< 0.1%)
 			if (votes.size () < 96 * 1024)
@@ -1226,9 +1226,9 @@ startup_time (std::chrono::steady_clock::now ())
 		logger.always_log ("Node ID: ", node_id.pub.to_account ());
 	}
 
-	const uint8_t * weight_buffer = network_params.is_live_network () ? nano_bootstrap_weights_live : nano_bootstrap_weights_beta;
-	size_t weight_size = network_params.is_live_network () ? nano_bootstrap_weights_live_size : nano_bootstrap_weights_beta_size;
-	if (network_params.is_live_network () || network_params.is_beta_network ())
+	const uint8_t * weight_buffer = network_params.network.is_live_network () ? nano_bootstrap_weights_live : nano_bootstrap_weights_beta;
+	size_t weight_size = network_params.network.is_live_network () ? nano_bootstrap_weights_live_size : nano_bootstrap_weights_beta_size;
+	if (network_params.network.is_live_network () || network_params.network.is_beta_network ())
 	{
 		nano::bufferstream weight_stream ((const uint8_t *)weight_buffer, weight_size);
 		nano::uint128_union block_height;
@@ -1467,7 +1467,7 @@ void nano::gap_cache::vote (std::shared_ptr<nano::vote> vote_a)
 				{
 					auto node_l (node.shared ());
 					auto now (std::chrono::steady_clock::now ());
-					node.alarm.add (node_l->network_params.is_test_network () ? now + std::chrono::milliseconds (5) : now + std::chrono::seconds (5), [node_l, hash]() {
+					node.alarm.add (node_l->network_params.network.is_test_network () ? now + std::chrono::milliseconds (5) : now + std::chrono::seconds (5), [node_l, hash]() {
 						auto transaction (node_l->store.tx_begin_read ());
 						if (!node_l->store.block_exists (transaction, hash))
 						{
@@ -1591,7 +1591,7 @@ void nano::node::keepalive_preconfigured (std::vector<std::string> const & peers
 {
 	for (auto i (peers_a.begin ()), n (peers_a.end ()); i != n; ++i)
 	{
-		keepalive (*i, network_params.default_node_port);
+		keepalive (*i, network_params.network.default_node_port);
 	}
 }
 
@@ -2099,7 +2099,7 @@ public:
 
 void nano::node::work_generate_blocking (nano::block & block_a)
 {
-	work_generate_blocking (block_a, network_params.publish_threshold);
+	work_generate_blocking (block_a, network_params.network.publish_threshold);
 }
 
 void nano::node::work_generate_blocking (nano::block & block_a, uint64_t difficulty_a)
@@ -2109,7 +2109,7 @@ void nano::node::work_generate_blocking (nano::block & block_a, uint64_t difficu
 
 void nano::node::work_generate (nano::uint256_union const & hash_a, std::function<void(uint64_t)> callback_a)
 {
-	work_generate (hash_a, callback_a, network_params.publish_threshold);
+	work_generate (hash_a, callback_a, network_params.network.publish_threshold);
 }
 
 void nano::node::work_generate (nano::uint256_union const & hash_a, std::function<void(uint64_t)> callback_a, uint64_t difficulty_a)
@@ -2120,7 +2120,7 @@ void nano::node::work_generate (nano::uint256_union const & hash_a, std::functio
 
 uint64_t nano::node::work_generate_blocking (nano::uint256_union const & block_a)
 {
-	return work_generate_blocking (block_a, network_params.publish_threshold);
+	return work_generate_blocking (block_a, network_params.network.publish_threshold);
 }
 
 uint64_t nano::node::work_generate_blocking (nano::uint256_union const & hash_a, uint64_t difficulty_a)
@@ -2631,7 +2631,7 @@ nano::election_vote_result nano::election::vote (nano::account rep, uint64_t seq
 	auto supply (node.online_reps.online_stake ());
 	auto weight (node.ledger.weight (transaction, rep));
 	auto should_process (false);
-	if (node.network_params.is_test_network () || weight > supply / 1000) // 0.1% or above
+	if (node.network_params.network.is_test_network () || weight > supply / 1000) // 0.1% or above
 	{
 		unsigned int cooldown;
 		if (weight < supply / 100) // 0.1% to 1%
@@ -2768,7 +2768,7 @@ void nano::active_transactions::confirm_frontiers (nano::transaction const & tra
 	~15 minutes for non-representative nodes, 3 minutes for representatives */
 	int representative_factor = representative ? 3 * 60 : 15 * 60;
 	// Decrease check time for test network
-	int test_network_factor = node.network_params.is_test_network () ? 1000 : 1;
+	int test_network_factor = node.network_params.network.is_test_network () ? 1000 : 1;
 	if (std::chrono::steady_clock::now () >= next_frontier_check)
 	{
 		size_t max_elections (max_broadcast_queue / 4);
@@ -2876,7 +2876,7 @@ void nano::active_transactions::request_confirm (std::unique_lock<std::mutex> & 
 				/* Escalation for long unconfirmed elections
 				Start new elections for previous block & source
 				if there are less than 100 active elections */
-				if (election_l->announcements % announcement_long == 1 && roots_size < 100 && !node.network_params.is_test_network ())
+				if (election_l->announcements % announcement_long == 1 && roots_size < 100 && !node.network_params.network.is_test_network ())
 				{
 					std::shared_ptr<nano::block> previous;
 					auto previous_hash (election_l->status.winner->previous ());
@@ -2949,7 +2949,7 @@ void nano::active_transactions::request_confirm (std::unique_lock<std::mutex> & 
 				if ((!rep_channels->empty () && node.rep_crawler.total_weight () > node.config.online_weight_minimum.number ()) || roots_size > 5)
 				{
 					// broadcast_confirm_req_base modifies reps, so we clone it once to avoid aliasing
-					if (!node.network_params.is_test_network ())
+					if (!node.network_params.network.is_test_network ())
 					{
 						if (confirm_req_bundle.size () < max_broadcast_queue)
 						{
@@ -2980,7 +2980,7 @@ void nano::active_transactions::request_confirm (std::unique_lock<std::mutex> & 
 				}
 				else
 				{
-					if (!node.network_params.is_test_network ())
+					if (!node.network_params.network.is_test_network ())
 					{
 						auto deque_l (node.network.udp_channels.random_set (100));
 						auto vec (std::make_shared<std::vector<std::shared_ptr<nano::transport::channel>>> ());
@@ -3020,7 +3020,7 @@ void nano::active_transactions::request_confirm (std::unique_lock<std::mutex> & 
 		node.network.flood_block_batch (rebroadcast_bundle);
 	}
 	// Batch confirmation request
-	if (!node.network_params.is_live_network () && !requests_bundle.empty ())
+	if (!node.network_params.network.is_live_network () && !requests_bundle.empty ())
 	{
 		node.network.broadcast_confirm_req_batch (requests_bundle, 50);
 	}
@@ -3068,7 +3068,7 @@ void nano::active_transactions::request_loop ()
 	{
 		request_confirm (lock);
 		const auto extra_delay (std::min (roots.size (), max_broadcast_queue) * node.network.broadcast_interval_ms * 2);
-		condition.wait_for (lock, std::chrono::milliseconds (node.network_params.request_interval_ms + extra_delay));
+		condition.wait_for (lock, std::chrono::milliseconds (node.network_params.network.request_interval_ms + extra_delay));
 	}
 }
 
@@ -3241,9 +3241,9 @@ void nano::active_transactions::adjust_difficulty (nano::block_hash const & hash
 		uint64_t average (static_cast<uint64_t> (sum / elections_list.size ()));
 		// Potential overflow check
 		uint64_t divider (1);
-		if (elections_list.size () > 1000000 && (average - node.network_params.publish_threshold) > elections_list.size ())
+		if (elections_list.size () > 1000000 && (average - node.network_params.network.publish_threshold) > elections_list.size ())
 		{
-			divider = ((average - node.network_params.publish_threshold) / elections_list.size ()) + 1;
+			divider = ((average - node.network_params.network.publish_threshold) / elections_list.size ()) + 1;
 		}
 		// Set adjusted difficulty
 		for (auto & item : elections_list)
