+++ nano/qt_system/entry.cpp
@@ -6,7 +6,7 @@
 
 int main (int argc, char ** argv)
 {
-	nano::network_params::set_active_network (nano::nano_networks::nano_test_network);
+	nano::network_constants::set_active_network (nano::nano_networks::nano_test_network);
 	QApplication application (argc, argv);
 	QCoreApplication::setOrganizationName ("Nano");
 	QCoreApplication::setOrganizationDomain ("nano.org");
