+++ nano/qt_test/entry.cpp
@@ -2,11 +2,14 @@
 #include <gtest/gtest.h>
 
 QApplication * test_application = nullptr;
-extern void force_nano_test_network ();
+namespace nano
+{
+void force_nano_test_network ();
+}
 
 int main (int argc, char ** argv)
 {
-	force_nano_test_network ();
+	nano::force_nano_test_network ();
 	QApplication application (argc, argv);
 	test_application = &application;
 	testing::InitGoogleTest (&argc, argv);
